Merged PR 746: Migrating strings to Resource generator
The PR adds the resource generator to the project which creates a strongly typed System.Strings class which is a replacement of partial SR class for referencing the resource members.

The SR utilities like SR.GetString and SR.Format have been left in the code.

Right now, to add a string, the user can modify the Strings.Resx file and then build the solution in VS to regenerate the Strings.Designer.cs to update it with the new resources.
TODO: find out how to generate the CS file from command prompt.

Related work items: #4625