Merged PR 769: Fix the resource manager used by SR.cs
The resource manager being used by SR.cs is not the one that loads the generated strings.
This is causing errors when we try to get strings using SR.GetString API.
Fixing the resource manager so that the debug asserts don't fail.

Not fixing this causes asserts in Debug configuration tests.