Merged PR 807: Allow OSGroup to override TargetsWindows
Right now the Directory.build.props figures out the environment Operating System.
This change checks if an OSGroup parameter is specified in the build command line, and if it is not, then the Directory.Build.Props is allowed to set it.
This is useful while doing Unix builds on Windows.