Merged PR 819: Netfx code
This pr brings in the code for netfx. I will make comments through out the pr where I had to make modifications in the code such as using SqlTypeWorkarounds.

Please note that for now, I have commented the Address.csproj and the test(SqlMetaDataTest) that uses it.

The netcore changes exist because I modified System.Data.SqlClient to Microsoft.Data.SqlClient in entire solution

Also, a lot of code from SafeNativeMethods and UnsadeNativeMethods was removed. It was because of compilation issues referencing OleDbHResult and was not referenced anywhere.