Merged PR 853: Check if the supplied resource is a resource ID or not.
Check if the supplied resource is a resource Id or not, this was found while importing the functional tests.
This PR fixes the cases where the Resource Id is supplied directly in the code, instead of using the Static properties from the ResXFileCodeGenerator generated file **Strings.Designer.cs**.
Given below is one of the places in the code where this issue would be observed. [[Link]](https://sqlclientdrivers.visualstudio.com/ADO.Net/_git/Microsoft.Data.SqlClient?path=%2Fsrc%2FMicrosoft.Data.SqlClient%2Fnetfx%2Fsrc%2FSystem%2FData%2FSqlClient%2FTdsParser.cs&version=GBmaster&line=1473&lineStyle=plain&lineEnd=1474&lineStartColumn=1&lineEndColumn=1)
The code change calls the **ResourceManager** again to check if the supplied resource is a Resource Id or not.
The **ResourceManager** returns null if the Resource Id is not valid, in which case the original value is retained, however if **ResourceManager** returns a Non-Null value, it means it is a Resource Id and the value of res is replaced by the retrieved Resource String.

Related work items: #5067