Merged PR 866: Added ability to read and set Assembly Version from command line during build
The Idea behind creating a separate target is that it doesn't use any additional dependencies/extensions.
This change instead of modifying original AssemblyInfo.cs file, creates a temporary copy of it during the build process, and pass it to the compiler. Generated assembly then has the modified version number, and original source file is left untouched.

The Assembly version can be supplied at run time using `/p:AssemblyFileVersion`
In case this property is not supplied at build it picks up the value from `<AssemblyFileVersion>` property in the `Directory.Build.props`
> MSBuild.exe .\Microsoft.Data.SqlClient.sln /p:Configuration=Debug /p:Platform=x86 **/p:AssemblyFileVersion=4.9.5.7**

This approach follows the tutorial mentioned at the link below:
[Original Page](http://www.lionhack.com/2014/02/13/msbuild-override-assembly-version/)
[Cached Page](https://web.archive.org/web/20180320135616/http://www.lionhack.com:80/2014/02/13/msbuild-override-assembly-version) incase the original doesn't work.

Related work items: #4868