Merged PR 904: 1st iteration of Always Encrypted Migration
Adding some of the fundamental classes that are required for the AE feature and some changes to the classes on which they depend. This is code only and no tests are included. They are coming in the next iteration. I wanted to keep this first submission as small as possible while keeping the main branch clean all tests passing and project compiling)