Merged PR 926: Remove ProviderFactory and DbConnectionStringBuilder
Remove ProviderFactory and DbConnectionStringBuilder

SqlConnectionStringBuilder could earlier override `ConvertValueToString` which was an internal method.
The override was being used to change value for the `Authentication` keyword to a more readable form.

Since we cannot override the internal, SqlClient will now have need to support the Enum names, in addition to existing supported value of Authentication Keyword.

Hence now Authentication = Sql Password can be specified as Authentication = SqlPassword.

If a user sets Authentication=Sql Password and then retrieves the connection string using
```
DbConnectionStringBuilder b = new SqlConnectionStringBuilder();
b.ConnectionString = "Authentication = Sql Password"
Console.WriteLine(b.ConnectionString);

```
The output would be
` Authentication="SqlPassword"`
To be able to consume this connection string and make it valid for consumption by the SqlConnectionStringBuilder, again, we need to add the support for the additional Values for Authentication Keyword.

The additional values are

| New Value  | Existing Equivalent   |
|---|---|
|  SqlPassword | Sql Password   |
| ActiveDirectoryPassword  | Active Directory Password  |
|  ActiveDirectoryIntegrated|  Active Directory Integrated |
|  ActiveDirectoryInteractive|Active Directory Interactive  |

I am awaiting support for netfx testing to be able to add tests for this change.