Merged PR 944: Revert "Remove DataPermission class and use reflection instead"
Reverting the changes for DbDataPermission. We cannot use reflection here because:
AddPermissionEntry internal method is not static, which means we have to create an instance of DbDataPermission class. However, this class is abstract and we won't be able to create an instance of it.

The original pr was causing tests to fail with error saying:
```
Unhandled Exception: System.TypeInitializationException: The type initializer for 'Microsoft.Data.SqlClient.SqlConnection' threw an exception. ---> System.NullReferenceException: Object reference not set to an instance of an object.
   at Microsoft.Data.SqlClient.SqlClientPermission.AddPermissionEntry(Object constr)
```