Merged PR 990: Changes to build system to use msbuild proj instead of scripts
The changes include a build.proj in the root folder.
The build.proj has all the important targets for build the tests and the product.

This allows reuse of a lot of existing artifacts instead of replicating the login in the build scripts.
Modified the build.ps1 to be a very thin layer to invoke the build.

Ideally the CI should be changed as well, however these changes are backward compatible.

I will modify the CI, once these changes are in so that the CI is more efficient.

The dev workflow on Windows, from the root folder with `build.proj`

`msbuild` - Builds NetFx and NetCore binaries with Debug Configuration (x64, x86, AnyCPU)
`msbuild /p:Configuration=Release` - Builds NetFx and NetCore binaries with Release Configuration (x64, x86, AnyCPU)
`msbuild /p:BuildNetFx=false` - To skip building .Net Framework on Windows
**Building Tests**
`msbuild /t:BuildTestsCore` - For .Net Core tests in Debug configuration
`msbuild /t:BuildTestsNetFx` - For .Net Framework tests Debug configuration

The above commands work on Unix as well. The tests and product for NetFx are not built, based on the OS Configuration state by MSBuild.

Also the failures and continue on errors of MSBuild can be incorporated here and no additional checks in scripts are needed.

On Unix, msbuild will be replaced with `dotnet msbuild`