Merged PR 994: msbuild /t:Clean should clean all obj folders
Current the Nuget Restore causes a lot of obj folders to be created in different project folders.
There is no straight-forward way that I found to unify the folders where the project.asset.json can be restored without conflicting overwrites with different projects.
I am enhancing the build Clean target to search for and delete all obj folders.
Tested and worked on Windows and Ubuntu.