Merged PR 1003: Reduce the Timeout for PoolBlockingPeriod Test
The Pool Blocking period tests are very slow since they wait for the failure to happen for the complete timeout period specified in the connection string.
Changing the tests to have a shorter timeout .

This should significantly reduce the execution time in our CI. Without the PoolBlockingPeriod enabled, the CI used to take 3-4 minutes per manual test leg. With PoolBlockingPeriod tests enabled, the CI takes much longer (about 12 minutes).

This should save us some of the waiting time in the CI due to PoolBlockingPeriod test.