Merged PR 1002: Remove unncessary project references from UDTs
None of the UDTs should depend on Microsoft.Data.SqlClient for netfx, for their creation.
The types needed to annotate UDTs come from System.Data
Removing the extra project reference from UDT projects. I believe they were added because we had brought in the Microsoft.SqlServer namespace into SCOOB. Since then, we have refactored code and the project reference is not needed in the UDT projects.

If we have to include these project references in the UDT projects, then this is a huge blocker, where the customers would have to rewrite and re-deploy UDT assemblies on Sql server, which is against the design goal.