Merged PR 1045: Port over the PR changes https://github.com/dotnet/corefx/pull/34047
Port over the PR changes from the PR https://github.com/dotnet/corefx/pull/34047

Relevant description from the PR
Many synchronous code paths contain lambda allocations for async usage which capture variables in function scope which cause the allocation and use of closure types to contain them. This set of changes attempts to prevent allocation of a closure by using a state tuple or by breaking the closure machinery into a separate method only called in the async case.

The CachedAsyncState field was being allocated in the sync code path, I have refined the logic slightly to avoid this unless it is needed.

I have introduced a couple of pre-boxed integer constants which prevent the repeated boxing of the default value, these are bof the invalid prepared handle value and the empty read packet.

I have added an `AsyncHelper.ContinueTaskWithState` function which uses the same logic as the existing `ContinueTask` and adds the ability to pass a state object to each of the delegate parameters, this helps to avoid closures in calling code.