Merged PR 1447: Port GH PR36678 Allow specifying a UDT SqlParameter value as bytes
Port change to both netcore and netfx.
https://github.com/dotnet/corefx/pull/36678

Adds special casing when writing UDT parameter values to the TDS stream so that if the value is byte[] or SqlBytes the value is sent to the server and not rejected as invalid. This allows users to handle serialization and deserialization logic without having to have SqlClient be aware of the types and without using inefficient text representations.

Related work items: #8819