Merged PR 1550: Port GitHub Corefx #35777
The initial PR was across the whole corefx project. I only bring in the changes from the common files used in our Scoob and from the files under System.Data.SqlClient originally. The changes are applied to NetFx as well if the same file exists.

Basically, most of the changes are replaceing Debug.Assert(false) with Debug.Fail and with proper string format.

PS. The README.MD command for building netcore test is out of date so sync it with the actually working on in the build.proj.

Related work items: #9114