Merged PR 1573: Adding missing AE code
In comparing API differences, I found that SqlBulkCopyOptions.AllowEncryptedValueModifications was not on the netcore side. Looking at the source file, it looked like the AE code was just not ported into it... but I'm wondering it there is other bulk copy related code that was not ported. If anyone has tried AE and bulk copy, let me know.