Fix | Fixes concurrent connection speed issues due to 'Authentication Context' cache not maintained in NetCore (#466)
Cache the authentication context in the connection pool on netcore

The authentication context cache of the connection pool was not being updated in the netcore code. This was resulting in poor connection speed for subsequent connections which used AAD auth since they could not take advantage of the cached token. This code looks to have been missed when AAD authentication was ported from netfx to netcore.