Reenable DataStreamTest.RunAllTestsForSingleServer_TCP, previously blocked by #5540
Most of the issues with it were debug-only tests that checked async behavior. Async read task was expected to throw, but was simply being waited, causing the test to fail. Fixed by replacing waits with checks for exception (as was used in a handful of similar tests).

One issue was due to a debug assert failing in the TDS state object. It's really buried in timeout code, so I have no idea what it's expected to do. However, the code immediately after it was flagged by my IDE as always being true due to the debug - why are we conditionally executing code if we expect it to always be a certain value? So, I removed the debug assert.

Also added a handful of todo notes to the test class because it needs breaking up. And it's full of smelly Thread.Wait code.