Cleanup | Remove unused UDT serialization code (#3462)
* Remove _isTopLevelUdt and its usage

This always holds the constant value of true. Code paths which rely on it being false will never be called.

* Remove unused property and fields

IsNullable and _padBuffer were only referenced when _isTopLevelUdt was false.
_nullInstance was only referenced from IsNullable

* Remove _skipNormalize

This was always set to false

* Remove unused parameter from constructor

* Remove needless call to GetFormat

* Remove unused fields

These fields are assigned to, but never used

* Remove unused parameter

SqlConnection.GetBytes output a Format parameter which was never used at its call sites

* Remove incorrect comment