Tests | Active Issues 1 (#3494)
* Remove AsyncTest

* Reenable DataStreamTest.RunAllTestsForSingleServer_TCP, previously blocked by #5540

Most of the issues with it were debug-only tests that checked async behavior. Async read task was expected to throw, but was simply being waited, causing the test to fail. Fixed by replacing waits with checks for exception (as was used in a handful of similar tests).

One issue was due to a debug assert failing in the TDS state object. It's really buried in timeout code, so I have no idea what it's expected to do. However, the code immediately after it was flagged by my IDE as always being true due to the debug - why are we conditionally executing code if we expect it to always be a certain value? So, I removed the debug assert.

Also added a handful of todo notes to the test class because it needs breaking up. And it's full of smelly Thread.Wait code.

* Inhibited tests that fail on Azure or Named Instances.

* Inhibited more tests that fail on Azure or Named Instances.

* Fixed test sensitive to row result ordering, and missing completeness checks.

* - Added asserts to boolean function calls.
- Updated loop to read XML results correctly.

* Fixed inaccurate test.

* GetUniqueNameForSqlServer() now includes the generated name in the exception if > 128 chars.

* Truncating generated server name if necessary, rather than failing the test.

* Fixed truncation to preserve closing ']' if present.

---------

Co-authored-by: Paul Medynski <31868385+paulmedynski@users.noreply.github.com>