Fix double deactivation in Transaction Cleanup.
Problem:
During execution of the test TestAutoEnlistment_TxScopeComplete and TestAutoEnlistment_TxScopeNonComplete in Debug mode,a debug assertion fails:
Method Debug.Fail failed with 'activated multiple times?

This occurs inside the DeactivateConnection method when returning a connection to the pool after a transaction completes.This assertion failure indicates that the internal _activateCount was already 0 when decrement was attempted again, deactivating the connection more than once.

Root Cause:
In "TransactionEnded" function,the connection used in a transaction is eventually returned to the general transaction pool via:
```Pool.PutObjectFromTransactedPool(transactedObject);```
The DeactivateConnection function is executed within this function which causes the _activateCount to be decremented a second time, resulting in a negative value and triggering the debug assertion. However,the connection had already been deactivated earlier as part of its transaction enlistment lifecycle
There is no subsequent ActivateConnection function for this DeactivateConnection function, which makes the _activateCount to to -1.

Fix:
We prevent the second DeactivateConnection to decrement the _activateCount counter, by passing the parameter "false" to DeactivateConnection.

cc: @uweigand @giritrivedi @saitama951