netcore: remove hardcoded SSL protocol flags
NativeProtocols enum values are aligned with the SslProtocol values.
The output is only ever used to return the SSL warning, which checks whether any of each relevant SSL protocol's bits are set. Normalization of the client and server bits are irrelevant for this purpose - the SSL warning tests for both.