Merge | TdsParserStateObjectNative (#3631)
* netfx, netcore: move duplicate WritePacketCache class

Duplicated in netcore by a class in TdsParserSafeHandles.Windows.cs. Remove the duplicate and add to netfx.

* netfx, netcore: sync comments and whitespace

* netcore: sync tracing

* netcore: remove unused methods

ReadAsyncCallback and WriteAsyncCallback only recursed.
Also sync whitespace in netfx.

* netfx: sync null coalescing operator usage

* netfx, netcore: cleanup SQLFallbackDNSCache call

Also, netfx: sync of AuthProviderInfo population to match netcore. We were assigning default values.

* netfx: remove unnecessary default values for CreatePhysicalSNIHandle parameters

* netcore, netfx: sync SNIHandle construction

Also sync debug assertion for SessionHandle

* netcore: remove hardcoded SSL protocol flags

NativeProtocols enum values are aligned with the SslProtocol values.
The output is only ever used to return the SSL warning, which checks whether any of each relevant SSL protocol's bits are set. Normalization of the client and server bits are irrelevant for this purpose - the SSL warning tests for both.

* netcore: sync timeouts of ReadSyncOverAsync

This is only ever called with a timeoutRemaining parameter value of GetTimeoutRemaining().

* netcore: sync server certificate filename marshalling

Send a null value to the native SNI by default, not an empty string

* Merge TdsParserStateObjectNative

* First code review response

Maintained comment in Dispose method post-merge

* Add explicit handling of SNI-layer SSL protocols

This guards against future changes to the SslProtocols enum